Retorno do INT para media
Atender os requisito do trabalho

+++ RegistroDeHumor.h
@@ -18,7 +18,7 @@ typedef struct registro
     char data[30];
     Humor humor;
     char motivo[100];
-    float notaDoDia;
+    int notaDoDia;
 } RegistroDeHumor;
 
 typedef struct nolista
